refactor: location 패키지 네임 region으로 변경
