+++ src/main/java/com/prography/lighton/common/exception/GlobalExceptionHandler.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 import com.prography.lighton.common.utils.ApiUtils;
 import com.prography.lighton.common.utils.ApiUtils.ApiResult;
-import com.prography.lighton.location.exception.NoSuchRegionException;
+import com.prography.lighton.region.exception.NoSuchRegionException;
 import com.prography.lighton.member.exception.InvalidMemberException;
 import com.prography.lighton.member.exception.NoSuchMemberException;
